services/weightd/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2
RPi.GPIO
paho-mqtt
orjson
python-multipart
